Refactoring symbol/label/arrow freq measurement into parent class
